ai-nl2sql-lite/api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
mlx>=0.19.0
mlx-lm>=0.19.0
psycopg2-binary>=2.9.0
sqlparse>=0.4.0
python-dotenv>=1.0.0
